pyproject.toml
[tool.poetry]
name = "ai-product-matching"
version = "0.1.0"
description = "Image-to-product matching pipeline using VLMs, Vector DBs, MongoDB, and Triton"
authors = ["Saran Jaya Thilak <saranjthilak@gmail.com>"]
license = "MIT"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.9,<3.11"
fastapi = "^0.110"
pymongo = "^4.13.2"
numpy = "^1.24"
faiss-cpu = "^1.7"
tritonclient = "^2.41"
gradio = "^4.25"
transformers = "^4.41.0"
python-dotenv = "^1.1.1"

tensorrt = {version = "*", optional = true}
pillow = "10.3.0"
requests = "^2.32.4"
sentence-transformers = "^5.0.0"

[tool.poetry.extras]
tensorrt = ["tensorrt"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
